Rest v2 v2 (#4)
* Move restv2 code to rest, fix point initializations from decimal fields, update github action

* Github action fix attempt #2

* Github action fix attempt #3

* Github action fix attempt #4

* Update UI to use new rest format
