+++ .github/workflows/pythonapp.yml
@@ -9,7 +9,7 @@ jobs:
     
     services:
       postgres:
-        image: postgres:12
+        image: postgis/postgis
         env:
           POSTGRES_DB: feedback_map_dev
           POSTGRES_USER: postgres
@@ -20,12 +20,13 @@ jobs:
 
     steps:
     - uses: actions/checkout@v1
-    - name: Set up Python 3.7
+    - name: Set up Python 3.11
       uses: actions/setup-python@v1
       with:
-        python-version: 3.7
+        python-version: 3.11
     - name: Install dependencies
       run: |
+        sudo apt-get install gdal-bin
         python -m pip install --upgrade pip
         pip install pipenv
         cd django_server; pipenv install --system --deploy
