--- django_server/feedback_map/restv2/serializers/__init__.py
@@ -1,6 +1,0 @@
-from .base import BaseMapDataPointSerializer
-from .map_data_point import MapDataPointCommentSerializer, MapDataPointCommentNotificationSerializer, \
-    DictMapDataPointSerializer, MapDataPointSerializer, TagSerializer
-from .password_reset import PasswordResetSerializer
-from .user import BaseUserSerializer, UserSerializer
-from .rounding_decimal_field import RoundingDecimalField
