--- django_server/feedback_map/restv2/tests/__init__.py
@@ -1,2 +1,0 @@
-from .rest_tests import *  # noqa
-from .map_data_points_tests import *
