--- django_server/feedback_map/restv2/views/__init__.py
@@ -1,3 +1,0 @@
-from .map_data_points import MapDataPointsViewSet, MapDataPointCommentsViewSet, \
-    MapDataPointCommentNotificationsViewSet, MapDataPointsGeoJSON, TagsViewSet
-from .recent_mappers import RecentMappersViewSet
