+++ django_server/feedback_map_config/settings.py
@@ -98,7 +98,7 @@ DATABASES = {
         "ENGINE": os.environ.get("SQL_ENGINE", "django.contrib.gis.db.backends.postgis"),
         "NAME": os.environ.get("POSTGRES_DB", "feedback_map_dev"),
         "USER": os.environ.get("POSTGRES_USER", "postgres"),
-        "PASSWORD": os.environ.get("POSTGRES_PASSWORD"),
+        "PASSWORD": os.environ.get("POSTGRES_PASSWORD", "feedback_map"),
         "HOST": os.environ.get("SQL_HOST", "localhost"),
         "PORT": os.environ.get("SQL_PORT", "5432"),
     }
